Reorganize tests
By default, the `tap` module reports a failure if any given test file
requires more than 30 seconds to execute. This limitation could be
extended through configuration, but declaring all tests in a single file
makes discovery difficult for new contributors.

Reorganize tests into files based on the feature under test, avoiding
the restriction of the `tap` module and supporting new contributors.
